test_requirements.txt
-r requirements.txt
pytest==4.0.2
pytest-cov==2.6.0
SATOSA==3.4.8
flake8==3.7.6
munch==2.3.2
requests-mock==1.5.2
PyYAML==3.13